store/tikv: resolve util dependencies (#22524)
